Move ForkInfo one directory level up
