+++ app/modules/topics/components/ForkInfo/index.js
@@ -9,9 +9,9 @@ import { TOPIC_EDITOR_ROUTE } from 'config/routes';
 import FetchWrapper from 'components/FetchWrapper';
 import makeRoute from 'lib/makeRoute';
 
-import actions from '../../../actions';
-import * as m from '../../../model';
-import selectors from '../../../selectors';
+import actions from '../../actions';
+import * as m from '../../model';
+import selectors from '../../selectors';
 
 type PassedProps = {|
   upstreamTopicId: string,
