+++ app/modules/topics/components/index.js
@@ -2,6 +2,7 @@
 
 import Course from './Course';
 import Editor from './Editor';
+import ForkInfo from './ForkInfo';
 import NewTopicCard from './NewTopicCard';
 import SlidesList from './SlidesList';
 import TopicInfo from './TopicInfo';
@@ -11,6 +12,7 @@ import Viewer from './Viewer';
 const index = {
   Course,
   Editor,
+  ForkInfo,
   NewTopicCard,
   SlidesList,
   TopicInfo,
